requirements.txt
fastapi
uvicorn
requests
pandas
numpy
python-dotenv
firebase-admin
google-cloud-firestore
google-cloud-storage
pydantic
